Yarn and NPM mode

+++ scripts/make-cache.sh
@@ -8,8 +8,18 @@ if ! [ -x "$(command -v zip)" ]; then
     apt-get update && apt-get install zip -y
 fi
 
-npm i -g yarn@latest
-yarn install --force
+if [ -f package-lock.json ]; then
+  # npm mode
+  npm i -g npm@latest
+  npm i
 
-sleep 3
-yarn bundle
+  sleep 3
+  npm run bundle
+else
+  # yarn mode
+  npm i -g yarn@latest
+  yarn install --force
+
+  sleep 3
+  yarn bundle
+fi
